Refactored update_batch_and_job_status to utils
