+++ src/shared/utils.py
@@ -31,3 +31,36 @@ def send_embeddings_to_webhook(embedded_chunks: list[dict], job):
     )
 
     return response
+
+
+def update_batch_and_job_status(job_id, batch_status, batch_id):
+    try:
+        if not job_id and batch_id:
+            job = safe_db_operation(batch_service.get_batch, batch_id)
+            job_id = job.job_id
+        updated_batch_status = safe_db_operation(batch_service.update_batch_status, batch_id, batch_status)
+        job = safe_db_operation(job_service.update_job_with_batch, job_id, updated_batch_status)
+        if job.job_status == JobStatus.COMPLETED:
+            logging.info(f"Job {job_id} completed successfully")
+        elif job.job_status == JobStatus.PARTIALLY_COMPLETED:
+            logging.info(f"Job {job_id} partially completed. {job.batches_succeeded} out of {job.total_batches} batches succeeded")
+        elif job.job_status == JobStatus.FAILED:
+            logging.info(f"Job {job_id} failed. {job.batches_succeeded} out of {job.total_batches} batches succeeded")
+                
+    except Exception as e:
+        logging.error('Error updating job and batch status: %s', e)
+        safe_db_operation(job_service.update_job_status, job_id, JobStatus.FAILED)    
+
+
+
+def update_batch_status(job_id, batch_status, batch_id, retries = None, bypass_retries=False):
+    try:
+        updated_batch_status = safe_db_operation(batch_service.update_batch_status, batch_id, batch_status)
+        logging.info(f"Status for batch {batch_id} as part of job {job_id} updated to {updated_batch_status}") 
+        if updated_batch_status == BatchStatus.FAILED and (retries == config.MAX_BATCH_RETRIES or bypass_retries):
+            logging.info(f"Batch {batch_id} failed. Updating job status.")
+            update_batch_and_job_status(job_id, BatchStatus.FAILED, batch_id)     
+    except Exception as e:
+        logging.error('Error updating batch status: %s', e)
+
+
